requirements.txt
sqlalchemy
pandas
python-dotenv
openpyxl
psycopg2-binary
pyarrow
fastparquet
minio
polars
apache-airflow-providers-postgres